requirements-dev.txt
factory-boy==3.3.*
tblib==3.*
coverage==7.*
django_debug_toolbar==4.4.*
isort==6.*
black==25.*
flake8==7.*
